split classes inside user and eventstore
