+++ cfxdb/eventstore/publication.py
@@ -10,9 +10,7 @@ import pprint
 import cbor
 import flatbuffers
 
-from .gen.eventstore import Event as EventGen
-from .gen.eventstore import Publication as PublicationGen
-from .gen.eventstore import Session as SessionGen
+from cfxdb.gen.eventstore import Publication as PublicationGen
 
 
 class _Publication(PublicationGen.Publication):
@@ -59,174 +57,6 @@ class _Publication(PublicationGen.Publication):
             _len = self._tab.VectorLen(o)
             return memoryview(self._tab.Bytes)[_off:_off + _len]
         return None
-
-
-class Session(object):
-    """
-    Persisted session database object.
-    """
-    def __init__(self, from_fbs=None):
-        self._from_fbs = from_fbs
-
-        self._session = None
-        self._joined_at = None
-        self._left_at = None
-        self._realm = None
-        self._authid = None
-        self._authrole = None
-
-    def marshal(self):
-        obj = {
-            'session': self.session,
-            'joined_at': self.joined_at,
-            'left_at': self.left_at,
-            'realm': self.realm,
-            'authid': self.authid,
-            'authrole': self.authrole,
-        }
-        return obj
-
-    def __str__(self):
-        return '\n{}\n'.format(pprint.pformat(self.marshal()))
-
-    @property
-    def session(self):
-        """
-        The WAMP session ID of the session.
-
-        :returns: session ID
-        :rtype: int
-        """
-        if self._session is None and self._from_fbs:
-            self._session = self._from_fbs.Session()
-        return self._session
-
-    @session.setter
-    def session(self, value):
-        assert type(value) == int
-        self._session = value
-
-    @property
-    def joined_at(self):
-        """
-        Timestamp when the session was joined by the router. Epoch time in ns.
-
-        :returns: Epoch time in ns when session joined
-        :rtype: int
-        """
-        if self._joined_at is None and self._from_fbs:
-            self._joined_at = self._from_fbs.JoinedAt()
-        return self._joined_at
-
-    @joined_at.setter
-    def joined_at(self, value):
-        assert type(value) == int
-        self._joined_at = value
-
-    @property
-    def left_at(self):
-        """
-        Timestamp when the session left the router. Epoch time in ns.
-
-        :returns: Epoch time in ns when session left - or 0 when session currently joined
-        :rtype: int
-        """
-        if self._left_at is None and self._from_fbs:
-            self._left_at = self._from_fbs.LeftAt()
-        return self._left_at
-
-    @left_at.setter
-    def left_at(self, value):
-        assert type(value) == int
-        self._left_at = value
-
-    @property
-    def realm(self):
-        """
-        The WAMP realm the session is/was joined on.
-
-        :returns: WAMP realm
-        :rtype: str
-        """
-        if self._realm is None and self._from_fbs:
-            self._realm = self._from_fbs.Realm().decode('utf8')
-        return self._realm
-
-    @realm.setter
-    def realm(self, value):
-        assert type(value) == str
-        self._realm = value
-
-    @property
-    def authid(self):
-        """
-        The WAMP authid the session was authenticated under.
-
-        :returns: WAMP authid
-        :rtype: str
-        """
-        if self._authid is None and self._from_fbs:
-            self._authid = self._from_fbs.Authid().decode('utf8')
-        return self._authid
-
-    @authid.setter
-    def authid(self, value):
-        assert type(value) == str
-        self._authid = value
-
-    @property
-    def authrole(self):
-        """
-        The WAMP authrole the session was authenticated under.
-
-        :returns: WAMP authrole
-        :rtype: str
-        """
-        if self._authrole is None and self._from_fbs:
-            self._authrole = self._from_fbs.Authrole().decode('utf8')
-        return self._authrole
-
-    @authrole.setter
-    def authrole(self, value):
-        assert type(value) == str
-        self._authrole = value
-
-    @staticmethod
-    def cast(buf):
-        return Session(SessionGen.Session.GetRootAsSession(buf, 0))
-
-    def build(self, builder):
-
-        realm = self.realm
-        if realm:
-            realm = builder.CreateString(realm)
-
-        authid = self.authid
-        if authid:
-            authid = builder.CreateString(authid)
-
-        authrole = self.authrole
-        if authrole:
-            authrole = builder.CreateString(authrole)
-
-        # now start and build a new object ..
-        SessionGen.SessionStart(builder)
-
-        SessionGen.SessionAddSession(builder, self.session)
-        SessionGen.SessionAddJoinedAt(builder, self.joined_at)
-        if self.left_at:
-            SessionGen.SessionAddLeftAt(builder, self.left_at)
-        if realm:
-            SessionGen.SessionAddRealm(builder, realm)
-        if authid:
-            SessionGen.SessionAddAuthid(builder, authid)
-        if authrole:
-            SessionGen.SessionAddAuthrole(builder, authrole)
-
-        # finish the object.
-        final = SessionGen.SessionEnd(builder)
-
-        return final
 
 
 class Publication(object):
@@ -846,160 +676,5 @@ class Publication(object):
 
         # finish the object.
         final = PublicationGen.PublicationEnd(builder)
-
-        return final
-
-
-class Event(object):
-    """
-    Persisted event database object.
-    """
-    def __init__(self, from_fbs=None):
-        self._from_fbs = from_fbs
-
-        self._timestamp = None
-        self._subscription = None
-        self._publication = None
-        self._receiver = None
-        self._retained = None
-        self._acknowledged_delivery = None
-
-    def marshal(self):
-        obj = {
-            'timestamp': self.timestamp,
-            'subscription': self.subscription,
-            'publication': self.publication,
-            'receiver': self.receiver,
-            'retained': self.retained,
-            'acknowledged_delivery': self.acknowledged_delivery,
-        }
-        return obj
-
-    def __str__(self):
-        return '\n{}\n'.format(pprint.pformat(self.marshal()))
-
-    @property
-    def timestamp(self):
-        """
-        Timestamp when the event was sent to the receiver. Epoch time in ns.
-
-        :returns: Epoc time in ns.
-        :rtype: int
-        """
-        if self._timestamp is None and self._from_fbs:
-            self._timestamp = self._from_fbs.Timestamp()
-        return self._timestamp
-
-    @timestamp.setter
-    def timestamp(self, value):
-        assert type(value) == int
-        self._timestamp = value
-
-    @property
-    def subscription(self):
-        """
-        The subscription ID this event is dispatched under.
-
-        :returns: The subscription ID.
-        :rtype: int
-        """
-        if self._subscription is None and self._from_fbs:
-            self._subscription = self._from_fbs.Subscription()
-        return self._subscription
-
-    @subscription.setter
-    def subscription(self, value):
-        assert type(value) == int
-        self._subscription = value
-
-    @property
-    def publication(self):
-        """
-        The publication ID of the dispatched event.
-
-        :returns: The publication ID.
-        :rtype: int
-        """
-        if self._publication is None and self._from_fbs:
-            self._publication = self._from_fbs.Publication()
-        return self._publication
-
-    @publication.setter
-    def publication(self, value):
-        assert type(value) == int
-        self._publication = value
-
-    @property
-    def receiver(self):
-        """
-        The WAMP session ID of the receiver.
-
-        :returns: The receiver ID.
-        :rtype: int
-        """
-        if self._receiver is None and self._from_fbs:
-            self._receiver = self._from_fbs.Receiver()
-        return self._receiver
-
-    @receiver.setter
-    def receiver(self, value):
-        assert type(value) == int
-        self._receiver = value
-
-    @property
-    def retained(self):
-        """
-        Whether the message was retained by the broker on the topic, rather than just published.
-
-        :returns: retained flag
-        :rtype: bool
-        """
-        if self._retained is None and self._from_fbs:
-            self._retained = self._from_fbs.Retained()
-        return self._retained
-
-    @retained.setter
-    def retained(self, value):
-        assert type(value) == bool
-        self._retained = value
-
-    @property
-    def acknowledged_delivery(self):
-        """
-        Whether this Event was to be acknowledged by the receiver.
-
-        :returns: acknowledged delivery flag
-        :rtype: bool
-        """
-        if self._acknowledged_delivery is None and self._from_fbs:
-            self._acknowledged_delivery = self._from_fbs.AcknowledgedDelivery()
-        return self._acknowledged_delivery
-
-    @acknowledged_delivery.setter
-    def acknowledged_delivery(self, value):
-        assert type(value) == bool
-        self._acknowledged_delivery = value
-
-    @staticmethod
-    def cast(buf):
-        return Event(EventGen.Event.GetRootAsEvent(buf, 0))
-
-    def build(self, builder):
-
-        # now start and build a new object ..
-        EventGen.EventStart(builder)
-
-        EventGen.EventAddTimestamp(builder, self.timestamp)
-        EventGen.EventAddSubscription(builder, self.subscription)
-        EventGen.EventAddPublication(builder, self.publication)
-        EventGen.EventAddReceiver(builder, self.receiver)
-
-        if self.retained is not None:
-            EventGen.EventAddRetained(builder, self.retained)
-        if self.acknowledged_delivery is not None:
-            EventGen.EventAddAcknowledgedDelivery(builder, self.acknowledged_delivery)
-
-        # finish the object.
-        final = EventGen.EventEnd(builder)
 
         return final
