+++ cfxdb/user/__init__.py
@@ -1,0 +1,20 @@
+from cfxdb.user.user import User
+from cfxdb.user.user_fbs import UserFbs
+from cfxdb.user.user_mrealm_role import UserMrealmRole
+from cfxdb.user.user_mrealm_role_fbs import UserMrealmRoleFbs
+from cfxdb.user.activation_token import ActivationToken
+from cfxdb.user.activation_token_fbs import ActivationTokenFbs
+from cfxdb.user.organization import Organization
+from cfxdb.user.organization_fbs import OrganizationFbs
+
+
+__all__ = (
+    'User',
+    'UserFbs',
+    'UserMrealmRole',
+    'UserMrealmRoleFbs',
+    'ActivationToken',
+    'ActivationTokenFbs',
+    'Organization',
+    'OrganizationFbs',
+)
