+++ cfxdb/user/user_mrealm_role_fbs.py
@@ -1,0 +1,6 @@
+class UserMrealmRoleFbs(object):
+    """
+    Database class for CFC user roles on a management realm using Flatbuffers.
+    """
+    def __init__(self):
+        raise NotImplementedError('not yet implemented')
